Combine both loca2 assets into one python file
